Support Apprenda 7.0.0 (#7)
* move install template to new path

* update vagrantfile and add references to k8s node

* add certs and k8s install/join scripts

* fix cloudid

* update readme

+++ scripts/ps/Download-Apprenda.ps1
@@ -1,10 +1,12 @@
 Add-Type -AssemblyName System.IO.Compression.FileSystem
 
-$url = "http://docs.apprenda.com/sites/default/files/Apprenda-6.8.0.zip"
+$url = "http://docs.apprenda.com/sites/default/files/Apprenda-7.0.0.zip"
 $output = "C:\Apprenda.zip"
-Write-Output "Downloading Apprenda, please wait..."
-(New-Object System.Net.WebClient).DownloadFile($url, $output)
-Write-Output "Download complete, unzipping Apprenda to C:\Apprenda..."
-[System.IO.Compression.ZipFile]::ExtractToDirectory("C:\Apprenda.zip", "C:\Apprenda")
-Write-Output "Now removing Apprenda zip..."
-Remove-Item "C:\Apprenda.zip"
+if (!(Test-Path -Path C:\Apprenda)) {
+  Write-Output "Downloading Apprenda, please wait..."
+  (New-Object System.Net.WebClient).DownloadFile($url, $output)
+  Write-Output "Download complete, unzipping Apprenda to C:\Apprenda..."
+  [System.IO.Compression.ZipFile]::ExtractToDirectory("C:\Apprenda.zip", "C:\Apprenda")
+  Write-Output "Now removing Apprenda zip..."
+  Remove-Item "C:\Apprenda.zip"
+}
